Add files via upload
From January through July of 2020, America tested 19% of its population.  8% of those tests were positive for Covid (diagnosis rate).  Therefore, the number of people diagnosed with covid relative to the total population in the country was 1.5%.  (Percent of Population with Covid)
By comparison, I created a hypothetical situation for other countries.  Were they to test 19% of their population like the US, in combination with that countries current diagnosis rate, what would there "Percent of Population with Covid" statistic be?

The below images show the "Percent of Population with Covid" for every country.  (For simplicity, each image is the countries of a particular continent)
Each image required a slight variation in the graph parameters of Code Part 2.